+++ CHANGELOG.md
@@ -41,10 +41,14 @@
 - [122](https://github.com/vegaprotocol/vegatools/pull/122) - Update protos for new position state message
 - [121](https://github.com/vegaprotocol/vegatools/pull/121) - Add option to `snapshotdb` to print a snapshot to a file as JSON
 - [140](https://github.com/vegaprotocol/vegatools/issue/140) - Add status column to liquidity commitment viewer
+- [156](https://github.com/vegaprotocol/vegatools/issue/156) - Upgraded go build versions to 1.16 + 1.17
+- [165](https://github.com/vegaprotocol/vegatools/issue/165) - Adding perftest subcommand
+- [184](https://github.com/vegaprotocol/vegatools/issue/184) - Refactoring perftest
 
 ### 🐛 Fixes
 - [78](https://github.com/vegaprotocol/vegatools/pull/78) - Fix build with missing dependency
 - [91](https://github.com/vegaprotocol/vegatools/pull/91) - Output of `snapshotdb` is now valid json
+- [173](https://github.com/vegaprotocol/vegatools/issues/173) - Add `MarketTracker` to checkpoint parser
 
 ## 0.41.1
 *2021-08-31*